+++ package.json
@@ -30,7 +30,7 @@
     "gray-matter": "^4.0.3",
     "lucide-react": "^0.383.0",
     "mini-svg-data-uri": "^1.4.4",
-    "next": "14.2.3",
+    "next": "^14.2.16",
     "next-themes": "^0.3.0",
     "react": "^18",
     "react-dom": "^18",
